1.嘉禾增加重要文件上传功能，且增加相应的权限， 2.嘉禾增加海外业务模块，且增加相应的权限， 3.利润分析报表错误修改

+++ FOS/src/main/java/haitai/fos/ffop/service/FCustomsDeclarationService.java
@@ -59,8 +59,8 @@ public class FCustomsDeclarationService {
 				}
 				consId = entity.getConsId();
 				num += Integer.parseInt(entity.getCudePackageNum());
-				weight += Integer.parseInt(entity.getCudeGrossWeight());
-				netWeight += Integer.parseInt(entity.getCudeNetWeight());
+				weight += Double.parseDouble(entity.getCudeGrossWeight());
+				netWeight += Double.parseDouble(entity.getCudeNetWeight());
 				idMap.put(oldId, entity.getCudeId());
 			}
 		}
